Refactor wallet identity into modules
Split wallet identity logic into focused submodules for public identities, backups, existing-wallet lookup, and identity set tracking

Extract xpub account-index derivation into a small XpubExt helper and reuse it in identity generation
