+++ rust/src/wallet_identity.rs
@@ -1,65 +1,21 @@
-use std::collections::HashSet;
-use std::str::FromStr as _;
+mod backup;
+mod existing;
+mod identity_set;
+mod public;
 
-use bdk_wallet::bitcoin::bip32::{ChildNumber, Fingerprint as BdkFingerprint, Xpub};
-use bdk_wallet::descriptor::ExtendedDescriptor;
-use cove_bdk::descriptor_ext::DescriptorExt as _;
-use cove_device::keychain::{Keychain, KeychainError};
-use cove_util::{result_ext::ResultExt as _, sha256_hash};
-use strum::IntoEnumIterator as _;
-use tracing::warn;
-
-use crate::backup::model::WalletBackup;
-use crate::database::{self, Database};
-use crate::keys::{self, Descriptors};
-use crate::network::Network;
-use crate::wallet::fingerprint::Fingerprint;
-use crate::wallet::{
-    WalletAddressType,
-    metadata::{WalletId, WalletMetadata, WalletMode, WalletType},
+pub(crate) use backup::identity_key_for_backup;
+pub(crate) use existing::{
+    collect_existing_wallet_identities, existing_public_wallet_by_identity_strict,
 };
+pub(crate) use identity_set::{ExistingWalletIdentitySet, WalletIdentityKey};
+pub(crate) use public::{PublicWalletIdentity, PublicWalletIdentityError};
 
-#[derive(Debug, Clone, Eq, PartialEq, Hash)]
-pub(crate) enum PublicWalletIdentity {
-    Descriptor(PublicDescriptorIdentity),
-    Xpub(PublicXpubIdentity),
+#[cfg(test)]
+pub(crate) mod test_support {
+    pub(crate) use super::identity_set::test_support::*;
 }
 
-#[derive(Debug, Clone, Eq, PartialEq, Hash)]
-pub(crate) struct PublicDescriptorIdentity {
-    external: CanonicalDescriptor,
-    internal: CanonicalDescriptor,
-}
-
-#[derive(Debug, Clone, Eq, PartialEq, Hash)]
-struct CanonicalDescriptor(String);
-
-impl CanonicalDescriptor {
-    fn from_descriptor(descriptor: &ExtendedDescriptor) -> Self {
-        Self(descriptor.to_normalized_string())
-    }
-}
-
-#[derive(Debug, Clone, Eq, PartialEq, Hash)]
-pub(crate) struct PublicXpubIdentity(String);
-
-#[derive(Debug, thiserror::Error)]
-pub(crate) enum PublicWalletIdentityError {
-    #[error("failed to read keychain public wallet material: {0}")]
-    Keychain(#[from] KeychainError),
-
-    #[error("failed to synthesize descriptor identity from xpub: {0}")]
-    Descriptor(#[from] keys::Error),
-
-    #[error("failed to parse fingerprint: {0}")]
-    Fingerprint(String),
-
-    #[error("failed to parse xpub: {0}")]
-    Xpub(String),
-
-    #[error("failed to parse descriptor: {0}")]
-    DescriptorParse(String),
-}
+use crate::{database, wallet::metadata::WalletId};
 
 #[derive(Debug, thiserror::Error)]
 pub(crate) enum WalletIdentityError {
@@ -79,445 +35,25 @@ pub(crate) enum WalletIdentityError {
     BackupXpub { wallet_name: String, source: PublicWalletIdentityError },
 }
 
-impl PublicWalletIdentity {
-    pub(crate) fn from_descriptors(descriptors: &Descriptors) -> Self {
-        Self::from_descriptor_pair(
-            &descriptors.external.extended_descriptor,
-            &descriptors.internal.extended_descriptor,
-        )
-    }
-
-    pub(crate) fn from_descriptor_pair(
-        external: &ExtendedDescriptor,
-        internal: &ExtendedDescriptor,
-    ) -> Self {
-        Self::Descriptor(PublicDescriptorIdentity {
-            external: CanonicalDescriptor::from_descriptor(external),
-            internal: CanonicalDescriptor::from_descriptor(internal),
-        })
-    }
-
-    pub(crate) fn from_descriptor_strs(
-        external: &str,
-        internal: &str,
-    ) -> Result<Self, PublicWalletIdentityError> {
-        let external = ExtendedDescriptor::from_str(external)
-            .map_err_str(PublicWalletIdentityError::DescriptorParse)?;
-        let internal = ExtendedDescriptor::from_str(internal)
-            .map_err_str(PublicWalletIdentityError::DescriptorParse)?;
-
-        Ok(Self::from_descriptor_pair(&external, &internal))
-    }
-
-    pub(crate) fn from_xpub(xpub: Xpub) -> Self {
-        Self::Xpub(PublicXpubIdentity(xpub.to_string()))
-    }
-
-    pub(crate) fn from_xpub_str_default_address_type(
-        xpub: &str,
-        fingerprint: Option<Fingerprint>,
-        network: Network,
-        address_type: WalletAddressType,
-    ) -> Result<Self, PublicWalletIdentityError> {
-        let xpub = Xpub::from_str(xpub).map_err_str(PublicWalletIdentityError::Xpub)?;
-
-        let Some(fingerprint) = fingerprint else {
-            return Ok(Self::from_xpub(xpub));
-        };
-
-        Self::from_xpub_default_address_type(xpub, fingerprint, network, address_type)
-    }
-
-    pub(crate) fn from_xpub_default_address_type(
-        xpub: Xpub,
-        fingerprint: Fingerprint,
-        network: Network,
-        address_type: WalletAddressType,
-    ) -> Result<Self, PublicWalletIdentityError> {
-        let coin_type = match network {
-            Network::Bitcoin => 0,
-            Network::Testnet | Network::Testnet4 | Network::Signet => 1,
-        };
-        let fingerprint = BdkFingerprint::from_str(&fingerprint.as_lowercase())
-            .map_err_str(PublicWalletIdentityError::Fingerprint)?;
-        let account = account_from_xpub(&xpub);
-        let descriptors = match address_type {
-            WalletAddressType::NativeSegwit => {
-                Descriptors::try_new_bip84(xpub, [84, coin_type, account], fingerprint)?
-            }
-            WalletAddressType::WrappedSegwit => {
-                Descriptors::try_new_bip49(xpub, [49, coin_type, account], fingerprint)?
-            }
-            WalletAddressType::Legacy => {
-                Descriptors::try_new_bip44(xpub, [44, coin_type, account], fingerprint)?
-            }
-        };
-
-        Ok(Self::from_descriptors(&descriptors))
-    }
-
-    pub(crate) fn from_existing_wallet(
-        metadata: &WalletMetadata,
-        keychain: &Keychain,
-    ) -> Result<Option<Self>, PublicWalletIdentityError> {
-        if let Some((external, internal)) = keychain.get_public_descriptor(&metadata.id)? {
-            return Ok(Some(Self::from_descriptor_pair(&external, &internal)));
-        }
-
-        let Some(xpub) = keychain.get_wallet_xpub(&metadata.id)? else {
-            return Ok(None);
-        };
-
-        if let Some(fingerprint) = metadata.master_fingerprint.as_deref()
-            && matches!(
-                metadata.wallet_type,
-                WalletType::Cold | WalletType::XpubOnly | WalletType::WatchOnly
-            )
-        {
-            return Self::from_xpub_default_address_type(
-                xpub,
-                *fingerprint,
-                metadata.network,
-                metadata.address_type,
-            )
-            .map(Some);
-        }
-
-        Ok(Some(Self::from_xpub(xpub)))
-    }
-
-    pub(crate) fn redacted_hash(&self) -> String {
-        let identity = match self {
-            Self::Descriptor(identity) => {
-                format!("descriptor:{}\n{}", identity.external.0, identity.internal.0)
-            }
-            Self::Xpub(identity) => format!("xpub:{}", identity.0),
-        };
-        let hash = sha256_hash(identity.as_bytes()).to_string();
-
-        hash[..16].to_string()
-    }
-}
-
-fn account_from_xpub(xpub: &Xpub) -> u32 {
-    match (xpub.depth, xpub.child_number) {
-        (3, ChildNumber::Hardened { index }) => index,
-        _ => 0,
-    }
-}
-
-#[derive(Debug, Clone, Eq, PartialEq, Hash)]
-pub(crate) enum WalletIdentityKey {
-    PublicIdentity {
-        identity: PublicWalletIdentity,
-        fingerprint: Option<Fingerprint>,
-        wallet_id: Option<WalletId>,
-        network: Network,
-        mode: WalletMode,
-    },
-    Fingerprint {
-        fingerprint: Fingerprint,
-        network: Network,
-        mode: WalletMode,
-    },
-    WalletId {
-        id: WalletId,
-        network: Network,
-        mode: WalletMode,
-    },
-}
-
-#[derive(Debug, Clone, Default)]
-pub(crate) struct ExistingWalletIdentitySet {
-    public_identities: HashSet<(PublicWalletIdentity, Network, WalletMode)>,
-    public_identity_fingerprints: HashSet<(Fingerprint, Network, WalletMode)>,
-    fingerprints: HashSet<(Fingerprint, Network, WalletMode)>,
-    wallet_ids: HashSet<(WalletId, Network, WalletMode)>,
-}
-
-impl ExistingWalletIdentitySet {
-    pub(crate) fn contains(&self, key: &WalletIdentityKey) -> bool {
-        match key {
-            WalletIdentityKey::PublicIdentity {
-                identity,
-                fingerprint,
-                wallet_id,
-                network,
-                mode,
-            } => {
-                self.public_identities.contains(&(identity.clone(), *network, *mode))
-                    || fingerprint.is_some_and(|fingerprint| {
-                        self.fingerprints.contains(&(fingerprint, *network, *mode))
-                    })
-                    || wallet_id
-                        .as_ref()
-                        .is_some_and(|id| self.wallet_ids.contains(&(id.clone(), *network, *mode)))
-            }
-            WalletIdentityKey::Fingerprint { fingerprint, network, mode } => {
-                self.fingerprints.contains(&(*fingerprint, *network, *mode))
-                    || self.public_identity_fingerprints.contains(&(*fingerprint, *network, *mode))
-            }
-            WalletIdentityKey::WalletId { id, network, mode } => {
-                self.wallet_ids.contains(&(id.clone(), *network, *mode))
-            }
-        }
-    }
-
-    pub(crate) fn insert(&mut self, key: WalletIdentityKey) {
-        match key {
-            WalletIdentityKey::PublicIdentity {
-                identity,
-                fingerprint,
-                wallet_id,
-                network,
-                mode,
-            } => {
-                self.public_identities.insert((identity, network, mode));
-
-                if let Some(fingerprint) = fingerprint {
-                    self.public_identity_fingerprints.insert((fingerprint, network, mode));
-                }
-
-                if let Some(wallet_id) = wallet_id {
-                    self.wallet_ids.insert((wallet_id, network, mode));
-                }
-            }
-            WalletIdentityKey::Fingerprint { fingerprint, network, mode } => {
-                self.fingerprints.insert((fingerprint, network, mode));
-            }
-            WalletIdentityKey::WalletId { id, network, mode } => {
-                self.wallet_ids.insert((id, network, mode));
-            }
-        }
-    }
-}
-
-pub(crate) fn identity_key_for_backup(
-    metadata: &WalletMetadata,
-    backup: &WalletBackup,
-) -> Result<WalletIdentityKey, WalletIdentityError> {
-    if metadata.wallet_type == WalletType::Hot
-        && let Some(fingerprint) = metadata.master_fingerprint.as_deref().copied()
-    {
-        return Ok(WalletIdentityKey::Fingerprint {
-            fingerprint,
-            network: metadata.network,
-            mode: metadata.wallet_mode,
-        });
-    }
-
-    if let Some(identity) = public_identity_from_backup(metadata, backup)? {
-        return Ok(WalletIdentityKey::PublicIdentity {
-            identity,
-            fingerprint: metadata.master_fingerprint.as_deref().copied(),
-            wallet_id: no_fingerprint_wallet_id(metadata),
-            network: metadata.network,
-            mode: metadata.wallet_mode,
-        });
-    }
-
-    if let Some(fingerprint) = metadata.master_fingerprint.as_deref().copied() {
-        return Ok(WalletIdentityKey::Fingerprint {
-            fingerprint,
-            network: metadata.network,
-            mode: metadata.wallet_mode,
-        });
-    }
-
-    Ok(WalletIdentityKey::WalletId {
-        id: metadata.id.clone(),
-        network: metadata.network,
-        mode: metadata.wallet_mode,
-    })
-}
-
-fn no_fingerprint_wallet_id(metadata: &WalletMetadata) -> Option<WalletId> {
-    metadata.master_fingerprint.is_none().then(|| metadata.id.clone())
-}
-
-pub(crate) fn collect_existing_wallet_identities()
--> Result<ExistingWalletIdentitySet, WalletIdentityError> {
-    let db = Database::global();
-    let keychain = Keychain::global();
-    let mut identities = ExistingWalletIdentitySet::default();
-
-    for network in Network::iter() {
-        for mode in [WalletMode::Main, WalletMode::Decoy] {
-            let wallets = db.wallets.get_all(network, mode)?;
-
-            for wallet in wallets {
-                let duplicate_key = existing_wallet_identity_key(wallet, keychain)?;
-                identities.insert(duplicate_key);
-            }
-        }
-    }
-
-    Ok(identities)
-}
-
-fn existing_wallet_identity_key(
-    metadata: WalletMetadata,
-    keychain: &Keychain,
-) -> Result<WalletIdentityKey, WalletIdentityError> {
-    if metadata.wallet_type != WalletType::Hot {
-        let identity =
-            PublicWalletIdentity::from_existing_wallet(&metadata, keychain).map_err(|source| {
-                WalletIdentityError::ExistingWalletPublicIdentity {
-                    wallet_id: metadata.id.clone(),
-                    source,
-                }
-            })?;
-
-        if let Some(identity) = identity {
-            return Ok(WalletIdentityKey::PublicIdentity {
-                identity,
-                fingerprint: metadata.master_fingerprint.as_deref().copied(),
-                wallet_id: no_fingerprint_wallet_id(&metadata),
-                network: metadata.network,
-                mode: metadata.wallet_mode,
-            });
-        }
-    }
-
-    if let Some(fingerprint) = metadata.master_fingerprint.as_deref().copied() {
-        return Ok(WalletIdentityKey::Fingerprint {
-            fingerprint,
-            network: metadata.network,
-            mode: metadata.wallet_mode,
-        });
-    }
-
-    Ok(WalletIdentityKey::WalletId {
-        id: metadata.id,
-        network: metadata.network,
-        mode: metadata.wallet_mode,
-    })
-}
-
-fn public_identity_from_backup(
-    metadata: &WalletMetadata,
-    backup: &WalletBackup,
-) -> Result<Option<PublicWalletIdentity>, WalletIdentityError> {
-    if let Some(descriptors) = &backup.descriptors {
-        let identity = PublicWalletIdentity::from_descriptor_strs(
-            &descriptors.external,
-            &descriptors.internal,
-        )
-        .map_err(|source| WalletIdentityError::BackupDescriptor {
-            wallet_name: metadata.name.clone(),
-            source,
-        })?;
-
-        return Ok(Some(identity));
-    }
-
-    if let Some(xpub) = &backup.xpub {
-        let fingerprint = metadata.master_fingerprint.as_deref().copied();
-        let identity = PublicWalletIdentity::from_xpub_str_default_address_type(
-            xpub,
-            fingerprint,
-            metadata.network,
-            metadata.address_type,
-        )
-        .map_err(|source| WalletIdentityError::BackupXpub {
-            wallet_name: metadata.name.clone(),
-            source,
-        })?;
-
-        return Ok(Some(identity));
-    }
-
-    Ok(None)
-}
-
-pub(crate) fn existing_public_wallet_by_identity_strict(
-    database: &Database,
-    keychain: &Keychain,
-    network: Network,
-    mode: WalletMode,
-    fingerprint: Fingerprint,
-    incoming_identity: &PublicWalletIdentity,
-) -> Result<Option<WalletMetadata>, WalletIdentityError> {
-    let wallets = database.wallets.get_all(network, mode)?;
-
-    matching_public_wallet_by_identity(wallets, keychain, fingerprint, incoming_identity, false)
-}
-
-fn matching_public_wallet_by_identity(
-    wallets: Vec<WalletMetadata>,
-    keychain: &Keychain,
-    fingerprint: Fingerprint,
-    incoming_identity: &PublicWalletIdentity,
-    allow_degraded_fingerprint_match: bool,
-) -> Result<Option<WalletMetadata>, WalletIdentityError> {
-    let mut degraded_same_fingerprint_wallet = None;
-
-    for wallet_metadata in wallets {
-        if !wallet_metadata.matches_fingerprint(fingerprint) {
-            continue;
-        }
-
-        let wallet_identity =
-            PublicWalletIdentity::from_existing_wallet(&wallet_metadata, keychain).map_err(
-                |source| WalletIdentityError::ExistingWalletPublicIdentity {
-                    wallet_id: wallet_metadata.id.clone(),
-                    source,
-                },
-            )?;
-
-        let Some(wallet_identity) = wallet_identity else {
-            degraded_same_fingerprint_wallet.get_or_insert(wallet_metadata);
-            continue;
-        };
-
-        if &wallet_identity == incoming_identity {
-            return Ok(Some(wallet_metadata));
-        }
-    }
-
-    let Some(wallet_metadata) = degraded_same_fingerprint_wallet else {
-        return Ok(None);
-    };
-
-    if !allow_degraded_fingerprint_match {
-        return Err(WalletIdentityError::MissingExistingWalletPublicIdentity {
-            wallet_id: wallet_metadata.id,
-        });
-    }
-
-    let wallet_id = wallet_metadata.id.clone();
-    let incoming_identity_hash = incoming_identity.redacted_hash();
-    warn!(
-        "same-fingerprint wallet missing public identity wallet_id={wallet_id} incoming_identity_hash={incoming_identity_hash}, falling back to fingerprint match"
-    );
-
-    Ok(Some(wallet_metadata))
-}
-
-#[cfg(test)]
-pub(crate) mod test_support {
-    use super::ExistingWalletIdentitySet;
-
-    pub(crate) trait ExistingWalletIdentitySetTestExt {
-        fn len(&self) -> usize;
-    }
-
-    impl ExistingWalletIdentitySetTestExt for ExistingWalletIdentitySet {
-        fn len(&self) -> usize {
-            self.public_identities.len() + self.fingerprints.len() + self.wallet_ids.len()
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use std::collections::HashMap;
+    use std::str::FromStr as _;
     use std::sync::{Arc, Once};
 
-    use cove_device::keychain::{KeychainAccess, KeychainError};
+    use bdk_wallet::bitcoin::bip32::{Fingerprint as BdkFingerprint, Xpub};
+    use cove_device::keychain::{Keychain, KeychainAccess, KeychainError};
 
+    use super::existing::matching_public_wallet_by_identity;
     use super::*;
-    use crate::backup::model::{DescriptorPair, WalletSecret};
+    use crate::backup::model::{DescriptorPair, WalletBackup, WalletSecret};
+    use crate::keys::Descriptors;
+    use crate::network::Network;
+    use crate::wallet::{
+        WalletAddressType,
+        fingerprint::Fingerprint,
+        metadata::{WalletMetadata, WalletType},
+    };
 
     #[derive(Debug, Default)]
     struct TestKeychain(parking_lot::Mutex<HashMap<String, String>>);
